Added Models tests

+++ backend/streamline/signals.py
@@ -12,18 +12,27 @@ CSV_PATH = settings.CSV_DIR
  
 @receiver(post_delete, sender=Url_PDF)
 def delete_pdf(sender, instance, **kwargs):
-    os.chdir(PDF_PATH)
-    print("--- PDF",instance.pdf_path,"DELETED --- ")
-    os.remove(instance.pdf_path)
+
+    if(os.path.isfile(instance.pdf_path)):
+
+        os.chdir(PDF_PATH)
+        print("--- PDF",instance.pdf_path,"DELETED --- ")
+        os.remove(instance.pdf_path)
 
 @receiver(post_delete, sender=Table_PDF)
 def delete_pdf_csv(sender, instance, **kwargs):
-    os.chdir(CSV_PATH)
-    print("--- CSV",instance.file_path,"DELETED --- ")
-    os.remove(instance.file_path)
+    
+    if(os.path.isfile(instance.file_path)):
+
+        os.chdir(CSV_PATH)
+        print("--- CSV",instance.file_path,"DELETED --- ")
+        os.remove(instance.file_path)
 
 @receiver(post_delete, sender=Table_HTML)
 def delete_html_csv(sender, instance, **kwargs):
-    os.chdir(CSV_PATH)
-    print("--- CSV",instance.file_path,"DELETED --- ")
-    os.remove(instance.file_path)
+
+    if(os.path.isfile(instance.file_path)):
+
+        os.chdir(CSV_PATH)
+        print("--- CSV",instance.file_path,"DELETED --- ")
+        os.remove(instance.file_path)
